+++ README.md
@@ -1,4 +1,4 @@
-# joogl
+# jgloom
 JGLOOM Java openGL: Object Oriented (man!) is Java library that wraps the functionality of the OpenGL API in a more
 object-oriented package. It puts OpenGL objects into Java interfaces, and their functions into static factories /
 container utils, allowing for code to be reactive to the user's OpenGL version/platform dependency.
