+++ src/main/java/jgloom/SharedLibraryLoader.java
@@ -1,4 +1,4 @@
-package joogl;
+package jgloom;
 
 import java.io.File;
 import java.io.FileInputStream;
