+++ src/main/java/jgloom/gl/GLBuffer.java
@@ -1,4 +1,4 @@
-package joogl.gl;
+package jgloom.gl;
 
 import org.lwjgl.opengl.GL43;
