+++ src/main/java/jgloom/gl/GLPlatform.java
@@ -1,4 +1,4 @@
-package joogl.gl;
+package jgloom.gl;
 
 /**
  * Signals the JOOGL library which platform the library is running on,
